Move actions_presenter
