+++ src/lib/y2storage/actions_presenter.rb
