+++ src/lib/y2storage/dialogs/proposal.rb
@@ -21,7 +21,7 @@
 
 require "yast"
 require "y2storage"
-require "y2storage/dialogs/actions_presenter"
+require "y2storage/actions_presenter"
 require "ui/installation_dialog"
 
 Yast.import "HTML"
